Creating file with worker handlers
